Registration and confirmation via email (#4)
* Registration process.. Still in progress

This commit includes improvements to the registration process, introducing the VerificationToken entity for email verification. It also incorporates updates to existing entities, including the addition of new fields to enhance functionality. Furthermore, a Builder class has been implemented to streamline the creation of User and Organisation objects.

* Complete registration process with email confirmation feature

Successfully set up JavaMailSender and implemented the logic for sending confirmation emails. However, there is a need for additional validations and improved error messages. Additionally, unit tests need to be added to ensure the robustness of the code. Kindly review the changes and provide your feedback so that we can proceed with merging it into the master branch.

* Small refactors

made some small refactors , deleted the JWTController and added AuthenticationController with proper urls.