MAJOR REFACTOR
Made major refactor , changed the id type from UUID to Long , since the Json parser was all the time upset and couldn't parse and exceptions were thrown. We will not rely on the ModelMapper , we will handle this manually. I already implemented Classes for this in the package factory - EntityFactory (we will build/create all the models/entities in this class) and the other is ResponseFactory(where we will build all the corresponding responses).