Service-Factory changes
deleted UserFactory and OrganisationFactory classes and moved the logic that was inside to EntityFactory. Also removed the field private Long organisationId from EventRequest , we will no longer need it , we will set the organisation of the event via token. Im extracting the logged user information that hit the endpoint and then i set this user that made the request to create an event , to the event itself.