Organisation details function (#25)
* Create organisation details controller

Created 2 controllers to show organisation details. One is for the admin , where conditions are not required (like if account is enabled , nonlocked , approved) - this endpoint/controller will be called in the admin html where we display all the organisations with the external library called (datatables). The other endpoint/controller covers conditions that i mentioned earlier. If the conditions are not met , we cannot load the details , and we will throw an exception and handle it properly. The endpoint to show details withoud conditions is located in AdminController , the other is in UnauthorizedUserController.

* Add event details

Added 2 queries that return event details , with as the same principle as the organisation details( with conditions and without).

* Fix code smells

Fixed code smells pointed out by Sonar.