Refactors on VerificationToken entity
Added three new fields on VerificationToken entity - LocalDateTime createdAt and updatedAt , and String type - either CONFIRM_EMAIL or FORGOTTEN_PASSWORD , because i included some check logic implementation when sending emails for confirmation accout ot password reset. I set the experation time to be 3 days (its subject to discussion) , also took care for appropriate exception scenarios. Everything is tested and working. And last but not least , added logic for generating new random password for the user that requested it. By dafault , the lenght of the new password is 15 (also subject to discussion) , it includes lower and upper letter cases , all the digits and "!@#$" symbols. Currently i assign them in a List<Character> , then shuffle the list with Collections and then Using Random class to append a char to String builder. So far eveyrting is working.