Security measures refactor
Implemented additional logic where if the token expiration date is invalid , i handle it with returning a proper message to the client. Also created a new scheduler , responsible for the token expiration date again. I make sure when the client close the browser , he looses the session , but the token remains valid , for this purpose i fetch all the "unexpired" tokens from the db , check if their expiration have expired and if , then i set the fields to expired(true). Runs every hour at the start of the hour.